requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dialogflow==1.1.0
Flask==2.0.1
google-api-core==1.30.0
google-auth==1.32.0
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
twilio==6.60.0
urllib3==1.26.5
Werkzeug==2.0.1
